requirements.txt
Django==2.0.2
django-appconf==1.0.2
django-compressor==2.2
django-tinymce==2.7.0
djangorestframework==3.7.7
Pillow==5.0.0
pytz==2018.3
rcssmin==1.0.6
rjsmin==1.0.12
